+++ .env
@@ -1,2 +1,2 @@
-APP_URL=http://192.168.2.135:8000
-API_URL=http://192.168.2.135:8000/api/v1/
+APP_URL=http://192.168.60.85:8000
+API_URL=http://192.168.60.85:8000/api/v1/
