+++ package.json
@@ -20,12 +20,15 @@
     "react-icons": "^5.4.0",
     "react-native": "0.76.4",
     "react-native-chart-kit": "^6.12.0",
+    "react-native-dotenv": "^3.4.11",
     "react-native-element-dropdown": "^2.12.2",
     "react-native-fs": "^2.20.0",
+    "react-native-gesture-handler": "^2.21.2",
     "react-native-keychain": "^9.2.2",
     "react-native-linear-gradient": "^2.8.3",
     "react-native-safe-area-context": "^5.0.0",
-    "react-native-screens": "^4.3.0",
+    "react-native-screens": "^4.4.0",
+    "react-native-splash-screen": "^3.3.0",
     "react-native-svg": "^15.10.1",
     "react-native-svg-transformer": "^1.5.0",
     "react-native-vector-icons": "^10.1.0",
