+++ screen/auth/AuthContext.js
@@ -1,0 +1,28 @@
+// src/context/AuthContext.js
+
+import React, {createContext, useState, useContext} from 'react';
+
+const AuthContext = createContext();
+
+export const useAuth = () => useContext(AuthContext);
+
+export const AuthProvider = ({children}) => {
+  const [user, setUser] = useState(null);
+  const [token, setToken] = useState(null);
+
+  const login = (userData, token) => {
+    setUser(userData);
+    setToken(token);
+  };
+
+  const logout = () => {
+    setUser(null);
+    setToken(null);
+  };
+
+  return (
+    <AuthContext.Provider value={{user, token, login, logout}}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
