+++ screen/components/splashscreen/SplashScreen.js
@@ -1,0 +1,40 @@
+import React, {useEffect} from 'react';
+import {View, Text, StyleSheet, ActivityIndicator} from 'react-native';
+
+const SplashScreen = ({navigation}) => {
+  useEffect(() => {
+    const checkLoginStatus = async () => {
+      try {
+        const credentials = await Keychain.getGenericPassword();
+        if (credentials) {
+          const token = credentials.password;
+          await fetchUser(token); // Ambil data user dengan token
+          navigation.replace('Home'); // Langsung masuk ke Home jika token valid
+        } else {
+          navigation.replace('Login'); // Navigasi ke Login jika tidak ada token
+        }
+      } catch (error) {
+        console.error('Error checking login status:', error);
+        navigation.replace('Login');
+      }
+    };
+    checkLoginStatus();
+  }, [navigation]);
+
+  return (
+    <View style={styles.container}>
+      <ActivityIndicator size="large" color="#0000ff" />
+      <Text>Loading...</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
+
+export default SplashScreen;
