+++ src/config/config.js
@@ -1,0 +1,6 @@
+// src/config.js
+const config = {
+  API_URL: 'http://192.168.60.85:8000/api/v1',
+};
+
+export default config;
